Updated the index.js and helper.js files
